+++ src/main/java/spring/backend/activity/presentation/ReadMonthlyActivityOverviewController.java
@@ -4,7 +4,6 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RestController;
 import spring.backend.activity.application.ReadMonthlyActivityOverviewService;
 import spring.backend.activity.dto.request.MonthlyActivityOverviewRequest;
