+++ src/main/java/spring/backend/activity/presentation/swagger/ReadActivitiesByMemberAndKeywordInMonthSwagger.java
@@ -2,10 +2,8 @@ package spring.backend.activity.presentation.swagger;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.ResponseEntity;
-import spring.backend.activity.domain.value.Keyword;
 import spring.backend.activity.dto.request.ActivitiesByMemberAndKeywordInMonthRequest;
 import spring.backend.activity.dto.response.ActivitiesByMemberAndKeywordInMonthResponse;
 import spring.backend.core.presentation.RestResponse;
