+++ src/main/java/spring/backend/auth/application/HandleOAuthLoginService.java
@@ -3,9 +3,9 @@ package spring.backend.auth.application;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Service;
-import spring.backend.auth.dto.response.LoginResponse;
-import spring.backend.auth.dto.response.OAuthAccessTokenResponse;
-import spring.backend.auth.dto.response.OAuthResourceResponse;
+import spring.backend.auth.presentation.dto.response.LoginResponse;
+import spring.backend.auth.presentation.dto.response.OAuthAccessTokenResponse;
+import spring.backend.auth.presentation.dto.response.OAuthResourceResponse;
 import spring.backend.auth.exception.AuthenticationErrorCode;
 import spring.backend.auth.infrastructure.OAuthRestClient;
 import spring.backend.auth.infrastructure.OAuthRestClientFactory;
