+++ src/main/java/spring/backend/auth/application/OnboardingSignUpService.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import spring.backend.auth.dto.request.OnboardingSignUpRequest;
-import spring.backend.auth.dto.response.OnboardingSignUpResponse;
+import spring.backend.auth.presentation.dto.request.OnboardingSignUpRequest;
+import spring.backend.auth.presentation.dto.response.OnboardingSignUpResponse;
 import spring.backend.auth.exception.AuthenticationErrorCode;
 import spring.backend.member.domain.entity.Member;
 import spring.backend.member.domain.repository.MemberRepository;
