+++ src/main/java/spring/backend/auth/infrastructure/kakao/KakaoOAuthRestClient.java
@@ -8,8 +8,8 @@ import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestClient;
 import org.springframework.web.util.UriComponentsBuilder;
-import spring.backend.auth.dto.response.OAuthAccessTokenResponse;
-import spring.backend.auth.dto.response.OAuthResourceResponse;
+import spring.backend.auth.presentation.dto.response.OAuthAccessTokenResponse;
+import spring.backend.auth.presentation.dto.response.OAuthResourceResponse;
 import spring.backend.auth.exception.AuthenticationErrorCode;
 import spring.backend.auth.infrastructure.OAuthRestClient;
 import spring.backend.auth.infrastructure.kakao.dto.KakaoResourceResponse;
