+++ src/main/java/spring/backend/auth/presentation/OnboardingSignUpController.java
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import spring.backend.auth.application.OnboardingSignUpService;
-import spring.backend.auth.dto.request.OnboardingSignUpRequest;
-import spring.backend.auth.dto.response.OnboardingSignUpResponse;
+import spring.backend.auth.presentation.dto.request.OnboardingSignUpRequest;
+import spring.backend.auth.presentation.dto.response.OnboardingSignUpResponse;
 import spring.backend.auth.presentation.swagger.OnboardingSignUpSwagger;
 import spring.backend.core.configuration.argumentresolver.LoginMember;
 import spring.backend.core.configuration.interceptor.Authorization;
