+++ src/main/java/spring/backend/auth/presentation/RotateAccessTokenController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import spring.backend.auth.application.RotateAccessTokenService;
-import spring.backend.auth.dto.response.RotateAccessTokenResponse;
+import spring.backend.auth.presentation.dto.response.RotateAccessTokenResponse;
 import spring.backend.core.presentation.RestResponse;
 
 @RestController
