+++ src/main/java/spring/backend/auth/presentation/dto/response/OnboardingSignUpResponse.java
@@ -1,4 +1,4 @@
-package spring.backend.auth.dto.response;
+package spring.backend.auth.presentation.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import spring.backend.member.domain.value.Gender;
