+++ src/main/java/spring/backend/auth/presentation/swagger/OnboardingSignUpSwagger.java
@@ -4,8 +4,8 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.ResponseEntity;
-import spring.backend.auth.dto.request.OnboardingSignUpRequest;
-import spring.backend.auth.dto.response.OnboardingSignUpResponse;
+import spring.backend.auth.presentation.dto.request.OnboardingSignUpRequest;
+import spring.backend.auth.presentation.dto.response.OnboardingSignUpResponse;
 import spring.backend.auth.exception.AuthenticationErrorCode;
 import spring.backend.core.configuration.swagger.ApiErrorCode;
 import spring.backend.core.exception.error.GlobalErrorCode;
