+++ src/test/java/spring/backend/auth/application/OnboardingSignUpServiceTest.java
@@ -6,8 +6,8 @@ import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
-import spring.backend.auth.dto.request.OnboardingSignUpRequest;
-import spring.backend.auth.dto.response.OnboardingSignUpResponse;
+import spring.backend.auth.presentation.dto.request.OnboardingSignUpRequest;
+import spring.backend.auth.presentation.dto.response.OnboardingSignUpResponse;
 import spring.backend.auth.exception.AuthenticationErrorCode;
 import spring.backend.core.exception.DomainException;
 import spring.backend.member.domain.entity.Member;
